crosscluster/physical: split PCR procs dest side instead
This allows the dest clsuter to choose to split less when it has fewer nodes.

Release note: none.
Epic: none.
