+++ pkg/ccl/crosscluster/physical/stream_ingestion_dist_test.go
@@ -522,3 +522,41 @@ func TestParallelInitialSplits(t *testing.T) {
 		require.Equal(t, sp.Key.String(), ts.mu.scatters[i].String())
 	}
 }
+
+func TestRepartition(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+
+	p := func(node, parts, start int) streamclient.PartitionInfo {
+		spans := make([]roachpb.Span, parts)
+		for i := range spans {
+			spans[i].Key = roachpb.Key(fmt.Sprintf("n%d-%d-a", node, i+start))
+			spans[i].EndKey = roachpb.Key(fmt.Sprintf("n%d-%d-b", node, i+start))
+		}
+		return streamclient.PartitionInfo{SrcInstanceID: node, Spans: spans}
+	}
+	for _, parts := range []int{1, 4, 100} {
+		for _, input := range [][]streamclient.PartitionInfo{
+			{p(1, 43, 0), p(2, 44, 0), p(3, 41, 0)},
+			{p(1, 1, 0), p(2, 1, 0), p(3, 1, 0)},
+			{p(1, 43, 0), p(2, 44, 0), p(3, 38, 0)},
+		} {
+			got := repartitionTopology(streamclient.Topology{Partitions: input}, parts)
+
+			var expectedSpans, gotSpans roachpb.Spans
+			for _, part := range input {
+				expectedSpans = append(expectedSpans, part.Spans...)
+			}
+			for _, part := range got.Partitions {
+				gotSpans = append(gotSpans, part.Spans...)
+			}
+			require.LessOrEqual(t, min(parts, len(input)), len(got.Partitions))
+			require.GreaterOrEqual(t, max(parts, len(input)), len(got.Partitions))
+
+			// Regardless of how we partitioned, make sure we have all the spans.
+			sort.Sort(expectedSpans)
+			sort.Sort(gotSpans)
+			require.Equal(t, len(expectedSpans), len(gotSpans))
+			require.Equal(t, expectedSpans, gotSpans)
+		}
+	}
+}
