+++ pkg/ccl/crosscluster/producer/replication_manager.go
@@ -190,13 +190,6 @@ var useStreaksInLDR = settings.RegisterBoolSetting(
 	false,
 )
 
-var ldrProcCount = settings.RegisterIntSetting(
-	settings.ApplicationLevel,
-	"logical_replication.producer.ingest_processor_parallelism",
-	"target number of stream partitions per source node",
-	1,
-)
-
 func (r *replicationStreamManagerImpl) PlanLogicalReplication(
 	ctx context.Context, req streampb.LogicalReplicationPlanRequest,
 ) (*streampb.ReplicationStreamSpec, error) {
@@ -225,8 +218,7 @@ func (r *replicationStreamManagerImpl) PlanLogicalReplication(
 		}
 	}
 
-	spec, err := buildReplicationStreamSpec(ctx, r.evalCtx, tenID, false, spans,
-		int(ldrProcCount.Get(&r.evalCtx.Settings.SV)), useStreaksInLDR.Get(&r.evalCtx.Settings.SV))
+	spec, err := buildReplicationStreamSpec(ctx, r.evalCtx, tenID, false, spans, useStreaksInLDR.Get(&r.evalCtx.Settings.SV))
 	if err != nil {
 		return nil, err
 	}
